--- tools/go_generics/generics_tests/all_stmts/opts.txt
@@ -1,1 +1,0 @@
--t=T=Q
