--- tools/go_generics/generics_tests/anon/opts.txt
@@ -1,1 +1,0 @@
--t=T=Q -suffix=New -anon
