--- tools/go_generics/generics_tests/consts/opts.txt
@@ -1,1 +1,0 @@
--c=c1=20 -c=z=600 -c=v=3.3 -c=s="def" -c=A=20 -c=C=100 -c=S="def" -c=T="ABC"
