--- tools/go_generics/generics_tests/imports/opts.txt
@@ -1,1 +1,0 @@
--t=T=sync.Mutex -c=n=math.Uint32 -c=m=math.Uint64 -import=sync=sync -import=math=mymathpath
