--- tools/go_generics/generics_tests/remove_typedef/opts.txt
@@ -1,1 +1,0 @@
--t=T=U
