--- tools/go_generics/generics_tests/simple/opts.txt
@@ -1,1 +1,0 @@
--t=T=Q -suffix=New
