+++ tools/go_generics/tests/all_stmts/input.go
