+++ tools/go_generics/tests/all_stmts/output.go
