+++ tools/go_generics/tests/all_types/lib/lib.go
