+++ tools/go_generics/tests/anon/input.go
