+++ tools/go_generics/tests/consts/input.go
