+++ tools/go_generics/tests/consts/output.go
