+++ tools/go_generics/tests/imports/input.go
