+++ tools/go_generics/tests/remove_typedef/input.go
