+++ tools/go_generics/tests/remove_typedef/output.go
