+++ tools/go_generics/tests/simple/input.go
