+++ tools/go_generics/tests/simple/output.go
